+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.1.3 / 2018-10-15
+==================
+
+  * Add `autoCleanUp` to configuration
+  * Add `excludeFields` to configuration
+  * Add `minimalMatch` as option for `toMatchSnapshot`
+  * Replace `json-normalize` dependency with own implementation
+
 0.1.2 / 2018-10-15
 ==================